requirements.txt
torch~=2.7.0
numpy~=1.26.4
matplotlib~=3.10.3
pillow~=10.4.0
ipython~=9.1.0
jinja2~=3.1.6
tqdm~=4.67.1
moviepy~=1.0.3
future~=1.0.0